requirements.txt
getmac==0.9.4
nudenet==3.0.8
Pillow==10.3.0
psutil==5.9.6
pycryptodome==3.20.0
pywin32==306
